backend/requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@master
asyncpg
jishaku
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
